# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/quest": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Quests"
                ],
                "summary": "Get many quests sorted by start time and finished status",
                "parameters": [
                    {
                        "maxLength": 3,
                        "minLength": 0,
                        "type": "array",
                        "items": {
                            "enum": [
                                "all",
                                "registered",
                                "owned"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Fields to return",
                        "name": "fields",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Number of quests to return for each field",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page ID to return. Mutually exclusive to multiple fields",
                        "name": "page_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quests.Quests"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Quests"
                ],
                "summary": "Create new quest",
                "parameters": [
                    {
                        "description": "Main quest information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.CreateQuestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.Quest"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "415": {
                        "description": "Unsupported Media Type"
                    }
                }
            }
        },
        "/quest/{id}/task-groups/bulk": {
            "patch": {
                "tags": [
                    "TaskGroups"
                ],
                "summary": "Patch task groups by creating new ones, delete, update and reorder all ones. Returns all exising task groups.",
                "parameters": [
                    {
                        "description": "Requests to delete/create/update task groups",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.TaskGroupsBulkUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.TaskGroup"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/quest/{quest_id}": {
            "get": {
                "tags": [
                    "Quests"
                ],
                "summary": "Get quest by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Quest ID",
                        "name": "quest_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.Quest"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Quests"
                ],
                "summary": "Update main quest information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Quest ID",
                        "name": "quest_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Quest information to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.UpdateQuestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.Quest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "415": {
                        "description": "Unsupported Media Type"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Quests"
                ],
                "summary": "Delete quest",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Quest ID",
                        "name": "quest_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "quests.PaginatedQuestsResponse": {
            "type": "object",
            "properties": {
                "next_page_id": {
                    "type": "string"
                },
                "quests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.Quest"
                    }
                }
            }
        },
        "quests.Quests": {
            "type": "object",
            "properties": {
                "all": {
                    "$ref": "#/definitions/quests.PaginatedQuestsResponse"
                },
                "owned": {
                    "$ref": "#/definitions/quests.PaginatedQuestsResponse"
                },
                "registered": {
                    "$ref": "#/definitions/quests.PaginatedQuestsResponse"
                }
            }
        },
        "storage.AccessType": {
            "type": "string",
            "enum": [
                "public",
                "link_only"
            ],
            "x-enum-varnames": [
                "AccessPublic",
                "AccessLinkOnly"
            ]
        },
        "storage.CreateQuestRequest": {
            "type": "object",
            "properties": {
                "access": {
                    "$ref": "#/definitions/storage.AccessType"
                },
                "description": {
                    "type": "string"
                },
                "finish_time": {
                    "type": "string"
                },
                "max_team_cap": {
                    "type": "integer"
                },
                "media_link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "registration_deadline": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "storage.CreateTaskGroupRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "order_idx": {
                    "type": "integer"
                },
                "pub_time": {
                    "type": "string"
                }
            }
        },
...
        "storage.DeleteTaskGroupRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "storage.Quest": {
            "type": "object",
            "properties": {
                "access": {
                    "$ref": "#/definitions/storage.AccessType"
                },
                "creator": {
                    "$ref": "#/definitions/storage.User"
                },
                "description": {
                    "type": "string"
                },
                "finish_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "max_team_cap": {
                    "type": "integer"
                },
                "media_link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "registration_deadline": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "storage.TaskGroup": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "orderIdx": {
                    "type": "integer"
                },
                "pubTime": {
                    "type": "string"
                },
                "quest": {
                    "$ref": "#/definitions/storage.Quest"
                }
            }
        },
        "storage.TaskGroupsBulkUpdateRequest": {
            "type": "object",
            "properties": {
                "create": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.CreateTaskGroupRequest"
                    }
                },
                "delete": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.DeleteTaskGroupRequest"
                    }
                },
                "update": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.UpdateTaskGroupRequest"
                    }
                }
            }
        },
...
        "storage.UpdateQuestRequest": {
            "type": "object",
            "properties": {
                "access": {
                    "$ref": "#/definitions/storage.AccessType"
                },
                "creator_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "finish_time": {
                    "type": "string"
                },
                "max_team_cap": {
                    "type": "integer"
                },
                "media_link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "registration_deadline": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "storage.UpdateTaskGroupRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "order_idx": {
                    "type": "integer"
                },
                "pub_time": {
                    "type": "string"
                }
            }
        },
        "storage.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}